+++ Assets/Scripts/Systems/MpSystem.cs
@@ -3,16 +3,16 @@ using UnityEngine;
 
 public class MpSystem : MonoBehaviour
 {
-	private int mp;
-	private int maxmp;
-	public int Mp {
+	private float mp;
+	private float maxmp;
+	public float Mp {
 		get { return mp; }
 		set {
 			MpChanged?.Invoke (value - mp);
 			mp = Mathf.Clamp (value, 0, maxmp); 
 		}
 	}
-	public int MaxMp {
+	public float MaxMp {
 		get { return maxmp; }
 		set {
 			MaxMpChanged?.Invoke(value - maxmp);
